size_ts to uint64s

+++ source/rshash.hpp
@@ -95,12 +95,12 @@ private:
     inline uint64_t find_minimiser(const uint64_t, const uint64_t, size_t &, size_t &, const uint64_t);
     inline void update_minimiser(const uint64_t, const uint64_t, uint64_t&, size_t &, size_t &, const uint64_t);
     inline bool check(const uint64_t, const uint64_t, uint64_t*, const size_t, const size_t, const size_t, const size_t, const uint64_t);
-    inline bool check_minimiser_pos(uint64_t *, const uint64_t, const uint64_t, const uint64_t, const size_t, const size_t, bool &, size_t &, size_t &, size_t &);
-    inline bool check_minimiser_pos2(uint64_t *, const uint64_t, const uint64_t, const uint64_t, const size_t, const size_t, const size_t, bool &, size_t &, size_t &, size_t &);
+    inline bool check_minimiser_pos(uint64_t *, const uint64_t, const uint64_t, const uint64_t, const size_t, const size_t, bool &, uint64_t &, uint64_t &, uint64_t &);
+    inline bool check_minimiser_pos2(uint64_t *, const uint64_t, const uint64_t, const uint64_t, const size_t, const size_t, const size_t, bool &, uint64_t &, uint64_t &, uint64_t &);
     inline bool check_overlap(uint64_t, uint64_t, uint64_t &, uint64_t &);
     inline void fill_buffer(uint64_t*, uint64_t*, size_t, size_t, const uint64_t, const uint64_t);
-    inline bool lookup_buffer(uint64_t*, uint64_t*, const size_t, const uint64_t, const uint64_t, size_t &, const size_t, const size_t, bool &, size_t &, size_t &);
-    inline bool extend_in_text(size_t&, size_t, size_t, bool, const uint64_t, const uint64_t, const uint64_t);
+    inline bool lookup_buffer(uint64_t*, uint64_t*, const size_t, const uint64_t, const uint64_t, uint64_t &, const size_t, const size_t, bool &, uint64_t &, uint64_t &);
+    inline bool extend_in_text(uint64_t&, uint64_t, uint64_t, bool, const uint64_t, const uint64_t, const uint64_t);
     const inline uint64_t get_word64(uint64_t pos);
     const inline uint64_t get_base(uint64_t pos);
 
@@ -144,14 +144,14 @@ private:
     template<int level>
     inline void fill_buffer(uint64_t *, uint64_t *, size_t, size_t, const uint64_t, const uint64_t);
     template<int level>
-    inline bool check_minimiser_pos(uint64_t *, const uint64_t, const uint64_t, const uint64_t, const size_t, const size_t, const size_t, bool &, size_t &, size_t &, size_t &);
+    inline bool check_minimiser_pos(uint64_t *, const uint64_t, const uint64_t, const uint64_t, const size_t, const size_t, const size_t, bool &, uint64_t &, uint64_t &, uint64_t &);
     template<int level>
-    inline bool check_minimiser_pos2(uint64_t *, const uint64_t, const uint64_t, const uint64_t, const size_t, const size_t, const size_t, const size_t, bool &, size_t &, size_t &, size_t &);
+    inline bool check_minimiser_pos2(uint64_t *, const uint64_t, const uint64_t, const uint64_t, const size_t, const size_t, const size_t, const size_t, bool &, uint64_t &, uint64_t &, uint64_t &);
     template<int level>
     inline bool check_overlap(uint64_t, uint64_t, uint64_t &, uint64_t &);
     template<int level>
-    inline bool lookup_buffer(uint64_t *, uint64_t *, const size_t, const uint64_t,  const uint64_t, size_t &, const size_t, const size_t, bool &, size_t &, size_t &);
-    inline bool extend_in_text(size_t&, size_t, size_t, bool, const uint64_t, const uint64_t);
+    inline bool lookup_buffer(uint64_t *, uint64_t *, const size_t, const uint64_t,  const uint64_t, uint64_t &, const size_t, const size_t, bool &, uint64_t &, uint64_t &);
+    inline bool extend_in_text(uint64_t&, uint64_t, uint64_t, bool, const uint64_t, const uint64_t);
     const inline uint64_t get_word64(uint64_t pos);
     const inline uint64_t get_base(uint64_t pos);
